⬆️  Move inmemory db to root level
