+++ elements/lisk-tree/src/error.ts
